Desafio 2 aula 05

+++ src/main/kotlin/br/com/alura/alugames/modelo/Plano.kt
@@ -1,8 +1,10 @@
 package br.com.alura.alugames.modelo
 
+import java.math.BigDecimal
+
 sealed class Plano(val tipo: String) {
 
-    open fun obterValor(aluguel: Aluguel): Double {
-        return aluguel.jogo.preco * aluguel.periodo.emDias
+    open fun obterValor(aluguel: Aluguel): BigDecimal {
+        return aluguel.jogo.preco * aluguel.periodo.emDias.toBigDecimal()
     }
 }
